requirements.txt
streamlit
pandas
pymupdf
openai
python-dotenv